requirements.txt
# 依赖清单
fastapi>=0.100.0
uvicorn[standard]
python-multipart
chromadb>=0.5.0
langchain-chroma>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0
pydantic>=2.0.0
Pillow>=10.0.0
